+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * progress bar for ui creation in delayed loading module
 * cdisc_dataset and dataset now return R6 class objects (RelationalDataset)
+* progress bar for ui creation in delayed loading module
+* new `RawDatasetConnector` to execute `CallableFunction` and obtain `RawDataset`
 * Change output of `keys` function to `keys` object. Updated checks in `DatasetConnector` and `dataset`.
 
 # teal 0.8.3